requirements.txt
Django==1.8.4
wheel==0.24.0
django-angular==0.7.15
distribute==0.7.3
pyBarcode==0.7
chemspipy==1.0.4
django-browserid==1.0.0
djangorestframework==3.2.3
django-filter==0.11.0
xkcd==2.3.1
pylatex==0.8.0
Pillow==3.2.0
# pytz>=2015.7
# selenium
# Would need pytractor, but not available in pip for python 3
# https://github.com/mwalkowski/pytractor
django_templated_email==0.5
easy_thumbnails==2.3
django-image-cropping==1.0.3
django-braces==1.9.0
